refactor(scrum-server): refactor user model
